+++ TriviaSpark.Web/Areas/Admin/Pages/Matches/Create.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model TriviaSpark.Web.Pages.MatchAdmin.CreateModel
+@model             CreateModel
 
 @{
     ViewData["Title"] = "Create";
