+++ TriviaSpark.Web/Areas/Admin/Pages/Matches/Create.cshtml.cs
@@ -1,11 +1,9 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.AspNetCore.Mvc.Rendering;
-using System;
-using System.Linq;
 using TriviaSpark.Web.Areas.Identity.Data;
 
-namespace TriviaSpark.Web.Pages.MatchAdmin
+namespace TriviaSpark.Web.Areas.Admin.Pages.Matches
 {
     public class CreateModel : PageModel
     {
