+++ TriviaSpark.Web/Areas/Admin/Pages/Matches/Delete.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model TriviaSpark.Web.Pages.MatchAdmin.DeleteModel
+@model             DeleteModel
 
 @{
     ViewData["Title"] = "Delete";
