+++ TriviaSpark.Web/Areas/Admin/Pages/Matches/Details.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model TriviaSpark.Web.Pages.MatchAdmin.DetailsModel
+@model             DetailsModel
 
 @{
     ViewData["Title"] = "Details";
