+++ TriviaSpark.Web/Areas/Admin/Pages/Matches/Details.cshtml.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
 using TriviaSpark.Web.Areas.Identity.Data;
 
-namespace TriviaSpark.Web.Pages.MatchAdmin
+namespace TriviaSpark.Web.Areas.Admin.Pages.Matches
 {
     public class DetailsModel : PageModel
     {
