+++ TriviaSpark.Web/Areas/Admin/Pages/Matches/Edit.cshtml
@@ -1,6 +1,6 @@
 ﻿@page
 @using TriviaSpark.Core.Match;
-@model TriviaSpark.Web.Pages.MatchAdmin.EditModel
+@model             EditModel
 
 @{
     ViewData["Title"] = "Edit";
