+++ TriviaSpark.Web/Areas/Admin/Pages/Matches/Edit.cshtml.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using TriviaSpark.Web.Areas.Identity.Data;
 
-namespace TriviaSpark.Web.Pages.MatchAdmin
+namespace TriviaSpark.Web.Areas.Admin.Pages.Matches
 {
     public class EditModel : PageModel
     {
