+++ TriviaSpark.Web/Areas/Admin/Pages/Matches/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model TriviaSpark.Web.Pages.MatchAdmin.IndexModel
+@model                                     Areas.Admin.Pages.Matches.IndexModel
 
 @{
     ViewData["Title"] = "Index";
