+++ TriviaSpark.Web/Areas/Admin/Pages/Questions/Delete.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model             DeleteModel
+@model                         DeleteModel
 
 @{
     ViewData["Title"] = "Delete";
