+++ TriviaSpark.Web/Areas/Admin/Pages/Questions/Details.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model DetailsModel
+@model             DetailsModel
 
 @{
     ViewData["Title"] = "Details";
