+++ TriviaSpark.Web/Areas/Admin/Pages/Questions/Edit.cshtml
@@ -1,11 +1,11 @@
 ﻿@page
-@model             CreateModel
+@model             EditModel
 
 @{
-    ViewData["Title"] = "Create";
+    ViewData["Title"] = "Edit";
 }
 
-<h1>Create</h1>
+<h1>Edit</h1>
 
 <h4>Question</h4>
 <hr />
@@ -13,11 +13,7 @@
     <div class="col-md-4">
         <form method="post">
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-            <div class="form-group">
-                <label asp-for="Question.QuestionId" class="control-label"></label>
-                <input asp-for="Question.QuestionId" class="form-control" />
-                <span asp-validation-for="Question.QuestionId" class="text-danger"></span>
-            </div>
+            <input type="hidden" asp-for="Question.QuestionId" />
             <div class="form-group">
                 <label asp-for="Question.QuestionText" class="control-label"></label>
                 <input asp-for="Question.QuestionText" class="form-control" />
@@ -44,24 +40,14 @@
                 <span asp-validation-for="Question.Source" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <label asp-for="Question.CreatedDate" class="control-label"></label>
-                <input asp-for="Question.CreatedDate" class="form-control" />
-                <span asp-validation-for="Question.CreatedDate" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="Question.ModifiedDate" class="control-label"></label>
-                <input asp-for="Question.ModifiedDate" class="form-control" />
-                <span asp-validation-for="Question.ModifiedDate" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <input type="submit" value="Create" class="btn btn-primary" />
+                <input type="submit" value="Save" class="btn btn-primary" />
             </div>
         </form>
     </div>
 </div>
 
 <div>
-    <a asp-page="Index">Back to List</a>
+    <a asp-page="./Index">Back to List</a>
 </div>
 
 @section Scripts {
