+++ TriviaSpark.Web/Areas/Admin/Pages/Questions/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model IndexModel
+@model                                     Areas.Admin.Pages.Questions.IndexModel
 
 @{
     ViewData["Title"] = "Index";
