+++ TriviaSpark.Web/Areas/Admin/Pages/Shared/_Layout.cshtml
@@ -1,0 +1,73 @@
+﻿@using TriviaSpark.Web.Status
+@using System.Reflection;
+@{
+    var currentYear = DateTime.Now.Year.ToString();
+    var status = new ApplicationStatus(Assembly.GetExecutingAssembly());
+}
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>@ViewData["Title"] - TriviaSpark.Web</title>
+
+    <link rel="stylesheet" href="/lib/bootstrap/dist/css/bootstrap.min.css" />
+    <link rel="stylesheet" href="/lib/datatables.net-bs/dataTables.bootstrap5.min.css" />
+    <link rel="stylesheet" href="~/lib/bootstrap/dist/css/bootstrap.min.css" />
+    <link rel="stylesheet" href="~/css/site.css" asp-append-version="true" />
+    <link rel="stylesheet" href="~/TriviaSpark.Web.styles.css" asp-append-version="true" />
+    @RenderSection("HeadScripts", required: false)
+</head>
+<body>
+    <header>
+        <nav class="navbar navbar-expand-sm navbar-toggleable-sm navbar-light bg-white border-bottom box-shadow mb-3">
+            <div class="container">
+                <a class="navbar-brand" asp-area="" asp-page="/Index">TriviaSpark.Web</a>
+                <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target=".navbar-collapse" aria-controls="navbarSupportedContent"
+                        aria-expanded="false" aria-label="Toggle navigation">
+                    <span class="navbar-toggler-icon"></span>
+                </button>
+                <div class="navbar-collapse collapse d-sm-inline-flex justify-content-between">
+                    <ul class="navbar-nav flex-grow-1">
+                        <li class="nav-item">
+                            <a class="nav-link text-dark" asp-area="" asp-page="/TriviaMatch">Trivia Match</a>
+                        </li>
+                    </ul>
+                    <partial name="_LoginPartial" />
+                </div>
+            </div>
+        </nav>
+    </header>
+    <div class="container">
+        <main role="main" class="pb-3">
+            @RenderBody()
+        </main>
+    </div>
+
+    <div class="text-center p-3" style="background-color: rgba(0, 0, 0, 0.2);">
+        <div class="container-fluid">
+            <div class="row">
+                <div class="col-lg-4 col-md-3 mb-4 float-start">
+                    <strong>Alpha</strong> Version: @status.BuildVersion.ToString()
+                </div>
+                <div class="col-lg-4 col-md-6 mb-4">
+                    &copy; @currentYear TriviaSpark.Web  -
+                    <a href="https://markhazleton.controlorigins.com" title="Mark Hazleton, Solutions Architect, Lifelong Learner, not sidetracked by sizzle" target="_blank">Mark Hazleton</a> -
+                    <a href="https://github.com/markhazleton" target="_blank">GitHub</a>
+                </div>
+                <div class="col-lg-4 col-md-3 mb-4 float-end">
+                    Last Built: @status.BuildDate.ToLocalTime().ToLongDateString() <br />
+                </div>
+            </div>
+        </div>
+    </div>
+
+
+    <script src="~/lib/jquery/jquery.js" asp-append-version="true"></script>
+    <script src="~/lib/bootstrap/dist/js/bootstrap.bundle.min.js" crossorigin="anonymous"></script>
+    <script src="~/lib/datatables/jquery.dataTables.min.js" asp-append-version="true"></script>
+    <script src="~/lib/datatables.net-bs/dataTables.bootstrap5.min.js"></script>
+    <script src="~/js/site.js" asp-append-version="true"></script>
+    @await RenderSectionAsync("Scripts", required: false)
+</body>
+</html>
