+++ TriviaSpark.Web/Areas/Admin/Pages/_ViewImports.cshtml
@@ -1,0 +1,4 @@
+﻿@using TriviaSpark.Web;
+@using TriviaSpark.Web.Areas.Admin.Pages;
+@namespace TriviaSpark.Web.Areas.Admin.Pages
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
