+++ TriviaSpark.Web/Properties/serviceDependencies.json
@@ -1,0 +1,9 @@
+{
+  "dependencies": {
+    "mssql1": {
+      "ignored": "true",
+      "type": "mssql",
+      "connectionId": "ConnectionStrings:TriviaSparkWebContextConnection"
+    }
+  }
+}
