+++ TriviaSpark.Web/TriviaSpark.Web.csproj
@@ -11,13 +11,13 @@
 		<PackageReadmeFile>README.md</PackageReadmeFile>
 		<RepositoryUrl>https://github.com/controlorigins/TriviaSpark</RepositoryUrl>
 		<RepositoryType>git</RepositoryType>
+		<UserSecretsId>3ab83c11-24d9-49af-bf8a-e04ace6159dc</UserSecretsId>
 	</PropertyGroup>
 	<ItemGroup>
 	  <PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="7.0.3" />
 	  <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="7.0.3" />
 	  <PackageReference Include="Microsoft.AspNetCore.Identity.UI" Version="7.0.3" />
 	  <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="7.0.3" />
-	  <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="7.0.3" />
 	  <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="7.0.3">
 	    <PrivateAssets>all</PrivateAssets>
 	    <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
@@ -36,5 +36,11 @@
 	  <None Update="TriviaSpark.Web.db">
 	    <CopyToOutputDirectory>Always</CopyToOutputDirectory>
 	  </None>
+	</ItemGroup>
+	<ItemGroup>
+	  <Folder Include="Areas\Admin\Controllers\" />
+	  <Folder Include="Areas\Admin\Data\" />
+	  <Folder Include="Areas\Admin\Models\" />
+	  <Folder Include="Areas\Admin\Views\" />
 	</ItemGroup>
 </Project>
