+++ TriviaSpark.Web/appsettings.Development.json
@@ -2,8 +2,8 @@
   "DetailedErrors": true,
   "Logging": {
     "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
+      "Default": "Error",
+      "Microsoft.AspNetCore": "Error"
     }
   }
 }
