+++ TriviaSpark.Web/nlog.config
@@ -1,9 +1,12 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
+<!-- 
+	From https://github.com/NLog/NLog/wiki/Getting-started-with-ASP.NET-Core-6 
+-->
 <nlog xmlns="http://www.nlog-project.org/schemas/NLog.xsd"
       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
       autoReload="true"
-      internalLogLevel="Info"
-      internalLogFile="c:\temp\internal-nlog-AspNetCore.txt">
+      internalLogLevel="Warn"
+      internalLogFile="c:\temp\internal-nlog-TriviaSpark.txt">
 
 	<!-- enable asp.net core layout renderers -->
 	<extensions>
@@ -13,11 +16,11 @@
 	<!-- the targets to write to -->
 	<targets>
 		<!-- File Target for all log messages with basic details -->
-		<target xsi:type="File" name="allfile" fileName="c:\temp\nlog-AspNetCore-all-${shortdate}.log"
+		<target xsi:type="File" name="allfile" fileName="c:\temp\nlog-TriviaSpark-all-${shortdate}.log"
 				layout="${longdate}|${event-properties:item=EventId:whenEmpty=0}|${level:uppercase=true}|${logger}|${message} ${exception:format=tostring}" />
 
 		<!-- File Target for own log messages with extra web details using some ASP.NET core renderers -->
-		<target xsi:type="File" name="ownFile-web" fileName="c:\temp\nlog-AspNetCore-own-${shortdate}.log"
+		<target xsi:type="File" name="ownFile-web" fileName="c:\temp\nlog-TriviaSpark-own-${shortdate}.log"
 				layout="${longdate}|${event-properties:item=EventId:whenEmpty=0}|${level:uppercase=true}|${logger}|${message} ${exception:format=tostring}|url: ${aspnet-request-url}|action: ${aspnet-mvc-action}|${callsite}" />
 
 		<!--Console Target for hosting lifetime messages to improve Docker / Visual Studio startup detection -->
@@ -27,15 +30,15 @@
 	<!-- rules to map from logger name to target -->
 	<rules>
 		<!--All logs, including from Microsoft-->
-		<logger name="*" minlevel="Trace" writeTo="allfile" />
+		<logger name="*" minlevel="Warn" writeTo="allfile" />
 
 		<!--Output hosting lifetime messages to console target for faster startup detection -->
-		<logger name="Microsoft.Hosting.Lifetime" minlevel="Info" writeTo="lifetimeConsole, ownFile-web" final="true" />
+		<logger name="Microsoft.Hosting.Lifetime" minlevel="Error" writeTo="lifetimeConsole, ownFile-web" final="true" />
 
 		<!--Skip non-critical Microsoft logs and so log only own logs (BlackHole) -->
-		<logger name="Microsoft.*" maxlevel="Info" final="true" />
-		<logger name="System.Net.Http.*" maxlevel="Info" final="true" />
+		<logger name="Microsoft.*" maxlevel="Error" final="true" />
+		<logger name="System.Net.Http.*" maxlevel="Warn" final="true" />
 
-		<logger name="*" minlevel="Trace" writeTo="ownFile-web" />
+		<logger name="*" minlevel="Warn" writeTo="ownFile-web" />
 	</rules>
 </nlog>
